+++ RELEASENOTES.md
@@ -8,6 +8,8 @@ This release includes the following changes since the
 [1.8.0-rc01 release](#180-rc01-2025-07-16):
 
 *   ExoPlayer:
+    *   Add getter for shuffle mode to the `ExoPlayer` interface
+        ([#2522](https://github.com/androidx/media/pull/2522)).
     *   More clearly throw an exception if `DefaultAudioSink` is accessed from
         multiple threads. If this happens due to a call to
         `RendererCapabilities.getFormatSupport` outside of the player, make sure